tests/api/requirements.txt
pytest
requests
python-dotenv
pytest-html
